+++ cmd/frontend/internal/graphqlbackend/access_tokens_test.go
@@ -16,13 +16,13 @@ import (
 
 // 🚨 SECURITY: This tests that users can't create tokens for users they aren't allowed to do so for.
 func TestMutation_CreateAccessToken(t *testing.T) {
-	mockAccessTokensCreate := func(t *testing.T, wantCreatorUserID int32) {
+	mockAccessTokensCreate := func(t *testing.T, wantCreatorUserID int32, wantScopes []string) {
 		db.Mocks.AccessTokens.Create = func(subjectUserID int32, scopes []string, note string, creatorUserID int32) (int64, string, error) {
 			if want := int32(1); subjectUserID != want {
 				t.Errorf("got %v, want %v", subjectUserID, want)
 			}
-			if want := []string{authz.ScopeUserAll}; !reflect.DeepEqual(scopes, want) {
-				t.Errorf("got %q, want %q", scopes, want)
+			if !reflect.DeepEqual(scopes, wantScopes) {
+				t.Errorf("got %q, want %q", scopes, wantScopes)
 			}
 			if want := "n"; note != want {
 				t.Errorf("got %q, want %q", note, want)
@@ -38,7 +38,7 @@ func TestMutation_CreateAccessToken(t *testing.T) {
 
 	t.Run("authenticated as user", func(t *testing.T) {
 		resetMocks()
-		mockAccessTokensCreate(t, 1)
+		mockAccessTokensCreate(t, 1, []string{authz.ScopeUserAll})
 		gqltesting.RunTests(t, []*gqltesting.Test{
 			{
 				Context: actor.WithActor(context.Background(), &actor.Actor{UID: 1}),
@@ -76,10 +76,63 @@ func TestMutation_CreateAccessToken(t *testing.T) {
 		}
 	})
 
+	t.Run("authenticated as user, using site-admin-only scopes", func(t *testing.T) {
+		resetMocks()
+		db.Mocks.Users.GetByCurrentAuthUser = func(ctx context.Context) (*types.User, error) {
+			return &types.User{ID: 1, SiteAdmin: false}, nil
+		}
+		defer func() { db.Mocks.Users.GetByCurrentAuthUser = nil }()
+
+		ctx := actor.WithActor(context.Background(), &actor.Actor{UID: 1})
+		result, err := (&schemaResolver{}).CreateAccessToken(ctx, &createAccessTokenInput{
+			User:   uid1GQLID,
+			Scopes: []string{authz.ScopeUserAll, authz.ScopeSiteAdminSudo},
+			Note:   "n",
+		})
+		if want := backend.ErrMustBeSiteAdmin; err != want {
+			t.Errorf("got err %v, want %v", err, want)
+		}
+		if result != nil {
+			t.Errorf("got result %v, want nil", result)
+		}
+	})
+
+	t.Run("authenticated as site admin, using site-admin-only scopes", func(t *testing.T) {
+		resetMocks()
+		mockAccessTokensCreate(t, 1, []string{authz.ScopeSiteAdminSudo, authz.ScopeUserAll})
+		db.Mocks.Users.GetByCurrentAuthUser = func(ctx context.Context) (*types.User, error) {
+			return &types.User{ID: 1, SiteAdmin: true}, nil
+		}
+		defer func() { db.Mocks.Users.GetByCurrentAuthUser = nil }()
+
+		gqltesting.RunTests(t, []*gqltesting.Test{
+			{
+				Context: actor.WithActor(context.Background(), &actor.Actor{UID: 1}),
+				Schema:  GraphQLSchema,
+				Query: `
+				mutation {
+					createAccessToken(user: "` + uid1GQLID + `", scopes: ["user:all", "site-admin:sudo"], note: "n") {
+						id
+						token
+					}
+				}
+			`,
+				ExpectedResult: `
+				{
+					"createAccessToken": {
+						"id": "QWNjZXNzVG9rZW46MQ==",
+						"token": "t"
+					}
+				}
+			`,
+			},
+		})
+	})
+
 	t.Run("authenticated as different user who is a site-admin", func(t *testing.T) {
 		resetMocks()
 		const differentSiteAdminUID = 234
-		mockAccessTokensCreate(t, differentSiteAdminUID)
+		mockAccessTokensCreate(t, differentSiteAdminUID, []string{authz.ScopeUserAll})
 		db.Mocks.Users.GetByCurrentAuthUser = func(ctx context.Context) (*types.User, error) {
 			return &types.User{ID: differentSiteAdminUID, SiteAdmin: true}, nil
 		}